fix: some major logical errors

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: iergin <iergin@student.42istanbul.com.t    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/22 07:18:21 by iergin            #+#    #+#             */
-/*   Updated: 2026/02/22 10:05:08 by iergin           ###   ########.fr       */
+/*   Updated: 2026/02/22 10:19:28 by iergin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,6 @@ int	ft_num_len(long n);
 int	hex_to_ptr(unsigned long n, int isBig);
 int	ft_print_ptr(void *p);
 int	ft_print_uint(unsigned int n);
-int	ft_print_hex(long n, int isBig);
+int	ft_print_hex(unsigned int n, int isBig);
 
 #endif
